refactor : 클래스의 각자 역할에 맞는 패키지로 변경
