+++ src/main/java/controller/command/CommandType.java
@@ -1,4 +1,4 @@
-package view;
+package controller.command;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
