+++ src/main/java/util/ExceptionMessages.java
@@ -1,4 +1,4 @@
-package domain.util;
+package util;
 
 public class ExceptionMessages {
     public static final String INVALID_DESTINATION = "장기말이 이동할 수 있는 경로가 아닙니다.";
