+++ src/main/java/util/MovablePointFinder.java
@@ -1,4 +1,4 @@
-package domain.util;
+package util;
 
 import domain.piece.Piece;
 import domain.point.Direction;
