+++ src/main/java/view/InputView.java
@@ -1,23 +1,22 @@
 package view;
 
-import controller.YNCommand;
+import controller.command.Command;
+import controller.command.YNCommand;
+
+import java.util.Scanner;
 
 public class InputView {
-    private final InputReader inputReader;
-
-    public InputView(InputReader inputReader) {
-        this.inputReader = inputReader;
-    }
+    private final Scanner scanner = new Scanner(System.in);
 
     public Command getGameCommand() {
-        return new Command(inputReader.readInput());
+        return new Command(scanner.nextLine());
     }
 
     public String getGameId() {
-        return inputReader.readInput();
+        return scanner.nextLine();
     }
 
     public YNCommand getYNCommand() {
-        return YNCommand.from(inputReader.readInput());
+        return YNCommand.from(scanner.nextLine());
     }
 }
