+++ src/test/java/dao/BoardDaoTest.java
@@ -3,7 +3,7 @@ package dao;
 import domain.point.File;
 import domain.point.Point;
 import domain.point.Rank;
-import domain.util.ExceptionMessages;
+import util.ExceptionMessages;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
