+++ src/test/java/domain/board/BoardTest.java
@@ -17,7 +17,7 @@ import domain.piece.queen.WhiteQueen;
 import domain.piece.rook.BlackRook;
 import domain.piece.rook.WhiteRook;
 import domain.point.Point;
-import domain.util.ExceptionMessages;
+import util.ExceptionMessages;
 import exception.CheckMateException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
