+++ src/test/java/domain/board/PawnTest.java
@@ -7,7 +7,7 @@ import domain.piece.pawn.WhitePawn;
 import domain.piece.rook.BlackRook;
 import domain.piece.rook.WhiteRook;
 import domain.point.Point;
-import domain.util.ExceptionMessages;
+import util.ExceptionMessages;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
